requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-rest-durin==1.1.0
djangorestframework==3.16.1
humanize==4.14.0
idna==3.11
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
